==> Step-4-TTS-Workflow.ipynb <==
## Notebook 4: TTS Workflow

We have the exact podcast transcripts ready now to generate our audio for the Podcast.

In this notebook, we will learn how to generate Audio using both [lucasnewman/f5-tts-mlx](https://huggingface.co/lucasnewman/f5-tts-mlx) models first. 

After that, we will use the output from Notebook 3 to generate our complete podcast

import IPython.display as ipd
from tqdm import tqdm
import requests



## Bringing it together: Making the Podcast

Okay now that we understand everything-we can now use the complete pipeline to generate the entire podcast

Let's load in our pickle file from earlier and proceed:

import pickle

with open('./resources/podcast_ready_data.pkl', 'rb') as file:
    PODCAST_TEXT = pickle.load(file)

We will concatenate the generated segments of audio and also their respective sampling rates since we will require this to generate the final audio

outputs = []

Function generate text for speaker 1

def generate_speaker1_audio_web(text, output_path):
    response = requests.post(url='http://localhost:8080/v1/audio/speech_gpt',json={'input':text,'speaker':'speaker1'})
    with open(output_path, 'wb') as file:
        file.write(response.content)

Function to generate text for speaker 2

def generate_speaker2_audio_web(text, output_path):
    response = requests.post(url='http://localhost:8080/v1/audio/speech_gpt',json={'input':text,'speaker':'speaker2'})
    with open(output_path, 'wb') as file:
        file.write(response.content)

Helper function to convert the numpy output from the models into audio

PODCAST_TEXT

Most of the times we argue in life that Data Structures isn't very useful. However, this time the knowledge comes in handy. 

We will take the string from the pickle file and load it in as a Tuple with the help of `ast.literal_eval()`

import ast

ast.literal_eval(PODCAST_TEXT)

#### Generating the Final Podcast

Finally, we can loop over the Tuple and use our helper functions to generate the audio

final_audio = None

i = 1

for item in tqdm(ast.literal_eval(PODCAST_TEXT), desc="Generating podcast segments", unit="segment"):
    speaker, text = item[0],item[1]
    output_path = f"./resources/segments/_podcast_segment_{i}.wav"
    if speaker == "Speaker 1":
        generate_speaker1_audio_web(text, output_path)
    else:  # Speaker 2
        generate_speaker2_audio_web(text, output_path)
    i += 1

# Combine the segments ./resources/segments
import re
import os
import numpy as np
import soundfile as sf

audio_files = sorted([f"./resources/segments/{file}" for file in os.listdir("./resources/segments")],
                     key=lambda x: int(re.search(r'segment_(\d+)\.wav', x).group(1)))

print("audio_files -> ", audio_files)
audio_data = []
for file in audio_files:
    data, rate = sf.read(file)
    audio_data.append(data)

audio_data = np.concatenate(audio_data)

### Output the Podcast

We can now save this as a wav file

SAMPLE_RATE = 32000
sf.write("./resources/_podcast.wav", audio_data, SAMPLE_RATE)

### Suggested Next Steps:

- Experiment with the prompts: Please feel free to experiment with the SYSTEM_PROMPT in the notebooks
- Extend workflow beyond two speakers
- Test other TTS Models
- Experiment with Speech Enhancer models as a step 5.

#fin
